pyproject.toml
[tool.poetry]
name = "spartid-ais"
version = "0.1.0"
description = "AIS example with data from Kystverket in Norway"
authors = ["Knut H Pederseen <knuthp@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
Flask = "^3.0.0"
requests = "^2.28.1"
flask-sqlalchemy = "^3.0.2"
flask-marshmallow = "^0.15.0"
pyais = "^2.5.7"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.8.0"
flask-migrate = "^4.0.5"
python-dotenv = "^1.0.0"
streamlit = "^1.28.2"
streamlit-folium = "^0.16.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
flake8 = "^6.0.0"
isort = "^5.10.1"
black = "^23.7.0"


[tool.poetry.group.datasciene.dependencies]
pandas = "^2.1.0"
jupyterlab = "^4.0.5"
pyarrow = "^14.0.1"


[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
